+++ src/Core/VRChatContentPublisher.Core/Settings/Models/AppSettings.cs
@@ -6,9 +6,12 @@ public sealed class AppSettings
 {
     public bool SkipFirstSetup { get; set; } = false;
     public bool UseRgbCyclingBackgroundMenu { get; set; } = false;
+
+    public bool NotificationsEnabled { get; set; } = true;
     public bool SendNotificationOnStartupSessionRestoreFailed { get; set; } = true;
     public bool SendNotificationOnTaskFailed { get; set; } = true;
     public bool SendNotificationOnPublicIpChanged { get; set; } = true;
+
     public string ConnectInstanceName { get; set; } = RandomWordsUtils.GetRandomWords();
     public int RpcServerPort { get; set; } = 59328;